src/requirements.txt
genanki==0.13.1
python-slugify==8.0.4
Requests==2.32.2
setuptools==70.0.0
